+++ cmd/todo_update_test.go
@@ -1,4 +1,0 @@
-package cmd
-
-// Tests for parseLink and isKnownLinkType have been moved to content_test.go
-// since those functions now live in todo_content.go