+++ internal/todo/integration/syncutil/footer.go
@@ -1,5 +1,0 @@
-package syncutil
-
-// SyncFooter is the attribution footer appended to issues synced to external
-// services (GitHub, ClickUp) so viewers know the issue is managed externally.
-const SyncFooter = `👾 *This issue is managed by [Jig](https://github.com/toba/jig). Edits made here will be overwritten.*`